:triangular_flag_on_post: Reorganised twitter bot directory structure
All global and local resources are linked to page objects

+++ bots/botsEntrypoint.fish
@@ -1,23 +1,39 @@
 #!/usr/bin/fish
 
 # AUTHOR      : avimehenwal
-# DATE        : 12-Jan-2020
-# PURPOSE     : entrypoint bots framework
+# DATE        : 15-April-2020
+# PURPOSE     : bot framework entrypoint
 #
-# Capture startup options and change to
-# change default configurations
+# This script runs the Avi Digital Marketting Bots
 
-kill -9 (pgrep firefox)
-kill -9 (pgrep chrome)
 
-clear
+set iter 5
+echo $iter
+
+if pgrep chrome
+    kill -9 (pgrep chrome)
+end
+if pgrep firefox
+    kill -9 (pgrep firefox)
+end
+
 robot \
     --dryrun \
+    # --include test-structure \
+    --exclude 'self-test'  \
+    --include twitter \
     --rpa \
-    --name 'My Bots' \
-    --outputdir ./bots/output \
-    --exclude skip \
+    --name "My Bots" \
+    # --suite BlessLikes \
+    --outputdir output \
+    --debugfile debugfile.log \
+    --loglevel DEBUG \
+    --reportbackground "#ADD8E6":"#FFCCCB" \
+    --suitestatlevel 3 \
+    --variablefile ../vars.yaml \
+    # --variable iter:$iter \
     ./
 
+# instagram-bot/instagram.robot
 
 # END
